MVP and Stretch Completed

+++ src/piggyBank/Coin.java
@@ -2,19 +2,19 @@ package piggyBank;
 
 public abstract class Coin {
     public abstract int getContent();
-    public abstract int getTotal();
-    public abstract int getValue();
-    public abstract void remove(int num);
+    public abstract double getTotal();
+    public abstract double getValue();
+    public abstract void remove(double num);
 
     //checks if the requested remove amount is greater than the existing total
-    public int check(int total){
-        int ttl=total;
+    public double check(double total){
+        double ttl=total;
 
         if(ttl>this.getTotal()){
             ttl-=this.getTotal();
             this.remove(this.getContent());
         }else{
-            int removed=ttl/this.getValue();
+            double removed=ttl/this.getValue();
             ttl-=removed*getValue();
             this.remove(removed);
         }
